refactor: 불필요한 클래스 제거 및 파일 분류 및 머지 시 발생하는 충돌 해결
