--- src/main/java/com/walletguardians/walletguardiansapi/domain/friend/entity/FriendStatusEnum.java
@@ -1,7 +1,0 @@
-package com.walletguardians.walletguardiansapi.domain.friend.entity;
-
-public enum FriendStatusEnum {
-    PENDING,
-    ACCEPTED,
-    REJECTED;
-}
