+++ src/main/java/com/walletguardians/walletguardiansapi/domain/friend/entity/status/FriendStatusEnum.java
@@ -1,0 +1,7 @@
+package com.walletguardians.walletguardiansapi.domain.friend.entity.status;
+
+public enum FriendStatusEnum {
+    PENDING,
+    ACCEPTED,
+    REJECTED;
+}
